requirements/base.txt
wheel
django
djangorestframework
psycopg2-binary
django-cors-headers
urllib3
uvicorn
django-environ
Pillow
django-oauth-toolkit
pytz
requests
django-simple-history
ujson
django-redis
django-tenant-schemas-q
jsonschema
drf-yasg
django-filter
pysftp

# JWT Authentication
django-rest-knox

# SSO Authentication
xmlsec
python3-saml
pandas
pdfkit
xlsxwriter
croniter
dateparser
pdf2image
PyPDF2

gevent
pymongo
rsa

pyjwt
djangorestframework-simplejwt

python-gnupg
regex

googlemaps
pycryptodome
openpyxl

boto3

pydantic
PyYAML